chore: prepare v1.52.0 release (#339)
* fix: address code review findings before v1.52.0 release

- Add package prefix to RejectSymlinkOutput error messages
- Document intentional race trade-off in pattern cache eviction
- Narrow MCP sanitizeError regex to filesystem paths only
- Change WithMaxFileSize signature from int64 to int for API consistency
- Use allowlist sanitization for discriminator JSON names


* chore: bump plugin version to 1.52.0

* chore: add benchmark results for v1.52.0


🤖 Generated automatically

---------

+++ parser/parser_options.go
@@ -294,12 +294,12 @@ func WithMaxCachedDocuments(count int) Option {
 // This prevents resource exhaustion from loading arbitrarily large files.
 // A value of 0 means use the default (10MB).
 // Returns an error if size is negative.
-func WithMaxFileSize(size int64) Option {
+func WithMaxFileSize(size int) Option {
 	return func(cfg *parseConfig) error {
 		if size < 0 {
 			return fmt.Errorf("parser: maxFileSize cannot be negative")
 		}
-		cfg.maxFileSize = size
+		cfg.maxFileSize = int64(size)
 		return nil
 	}
 }
